pyproject.toml
[project]
name = "pdf-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.6",
    "passlib[bcrypt]>=1.7.4",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sentence-transformers>=5.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

[tool.ruff.lint.per-file-ignores]
"**/*.py" = ["B008"]
